remove deadcode
