+++ pkg/controller/broadcastjob/broadcastjob_controller.go
@@ -44,11 +44,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 )
-
-var log = logf.Log.WithName("controller")
 
 var controllerKind = appsv1alpha1.SchemeGroupVersion.WithKind("BroadcastJob")
