+++ pkg/controller/statefulset/stateful_set_utils.go
@@ -36,28 +36,7 @@ import (
 	"k8s.io/kubernetes/pkg/controller/history"
 )
 
-// maxUpdateRetries is the maximum number of retries used for update conflict resolution prior to failure
-const maxUpdateRetries = 10
-
-// errUpdateConflict is the error used to indicate that the maximum number of retries against the API server have
-// been attempted and we need to back off
-var errUpdateConflict = fmt.Errorf("aborting update after %d attempts", maxUpdateRetries)
 var patchCodec = scheme.Codecs.LegacyCodec(appsv1alpha1.SchemeGroupVersion)
-
-// overlappingStatefulSets sorts a list of StatefulSets by creation timestamp, using their names as a tie breaker.
-// Generally used to tie break between StatefulSets that have overlapping selectors.
-type overlappingStatefulSets []*appsv1alpha1.StatefulSet
-
-func (o overlappingStatefulSets) Len() int { return len(o) }
-
-func (o overlappingStatefulSets) Swap(i, j int) { o[i], o[j] = o[j], o[i] }
-
-func (o overlappingStatefulSets) Less(i, j int) bool {
-	if o[i].CreationTimestamp.Equal(&o[j].CreationTimestamp) {
-		return o[i].Name < o[j].Name
-	}
-	return o[i].CreationTimestamp.Before(&o[j].CreationTimestamp)
-}
 
 // statefulPodRegex is a regular expression that extracts the parent StatefulSet and ordinal from the Name of a Pod
 var statefulPodRegex = regexp.MustCompile("(.*)-([0-9]+)$")
