+++ pkg/controller/statefulset/statefulset_controller.go
@@ -45,15 +45,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
 	"sigs.k8s.io/controller-runtime/pkg/source"
-)
-
-var log = logf.Log.WithName("controller")
-
-const (
-	// period to relist statefulsets and verify pets
-	statefulSetResyncPeriod = 30 * time.Second
 )
 
 // controllerKind contains the schema.GroupVersionKind for this controller type.
